refactor(tools/versioning_automation): share open PR issue reference lookup
