+++ tools/versioning_automation/src/tests/mod.rs
@@ -5,6 +5,7 @@ mod cli_action;
 mod compare_snapshot;
 mod gh_cli;
 mod git_cli;
+mod open_pr_issue_refs;
 mod pr_remote_snapshot;
 mod pr_run_snapshot;
 mod repo_name;
